Update README.md
In this capstone, we will predict if the Falcon 9 first stage will land successfully. SpaceX advertises Falcon 9 rocket launches on its website, with a cost of 62 million dollars; other providers cost upward of 165 million dollars each, much of the savings is because SpaceX can reuse the first stage. Therefore if we can determine if the first stage will land, we can determine the cost of a launch. This information can be used if an alternate company wants to bid against SpaceX for a rocket launch. In this module, you will be provided with an overview of the problem and the tools you need to complete the course.
Learning Objectives
Develop Python code to manipulate data in a Pandas data frame
Convert a JSON file into a Create a Python Pandas data frame by converting a JSON file
Create a Jupyter notebook and make it sharable using GitHub
Utilize data science methodologies to define and formulate a real-world business problem
Utilize your data analysis tools to load a dataset, clean it, and find out interesting insights from it